requirements.txt
selenium>=4.0,<5.0
pytest>=8.0,<9.0
allure-pytest>=2.13,<3.0
webdriver-manager>=4.0,<5.0
